Separating des probabilistic into several files
