+++ deslib/des/probabilistic/__init__.py
@@ -1,0 +1,14 @@
+from .base import BaseProbabilistic
+from .deskl import DESKL
+from .exponential import Exponential
+from .logarithmic import Logarithmic
+from .minimum_difference import MinimumDifference
+from .rrc import RRC
+
+
+__all__ = ['BaseProbabilistic',
+           'DESKL',
+           'Exponential',
+           'Logarithmic',
+           'MinimumDifference',
+           'RRC']
